Give redux and action sagas more info via payloads
This is how we make a thing we can trust. We can select stuff from the
state by using `select` effect. We can’t trust it though because while
we’re processing one action, another can alter the state.

If we pass the information via action’s payload we don’t rely on the
mutable state. One approach to “trustless” it to rely on yourself. ;)